requirements.txt
Flask==2.3.2
pandas==2.1.1
scikit-learn==1.3.1
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
